% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BatteryMCMC-package.R
\docType{package}
\name{BatteryMCMC-package}
\alias{BatteryMCMC}
\alias{BatteryMCMC-package}
\title{BatteryMCMC}
\description{
...
}
\details{
...
}
\author{
Pierre E. Jacob <pierre.jacob.work@gmail.com>
}
\keyword{package}

